+++ lib/fpath_base/src/fpath_base.ml
@@ -3,7 +3,6 @@ module Fpath = struct
   include T
   include Comparable.Make (T)
 
-  let hash t = String.hash (T.to_string t)
   let hash_fold_t state t = String.hash_fold_t state (T.to_string t)
 end
 
@@ -12,7 +11,6 @@ module Absolute_path = struct
   include T
   include Comparable.Make (T)
 
-  let hash t = String.hash (T.to_string t)
   let hash_fold_t state t = String.hash_fold_t state (T.to_string t)
 end
 
@@ -21,15 +19,15 @@ module Relative_path = struct
   include T
   include Comparable.Make (T)
 
-  let hash t = String.hash (T.to_string t)
   let hash_fold_t state t = String.hash_fold_t state (T.to_string t)
 end
 
-module Fpart = struct
-  module T = Fpath_sexp0.Fpart
+module Fsegment = struct
+  module T = Fpath_sexp0.Fsegment
   include T
   include Comparable.Make (T)
 
-  let hash t = String.hash (T.to_string t)
   let hash_fold_t state t = String.hash_fold_t state (T.to_string t)
 end
+
+module Fpart = Fsegment