+++ lib/fpath_sexp0/src/fpath_sexp0.ml
@@ -6,4 +6,5 @@ end
 
 module Absolute_path = Path.Absolute_path
 module Relative_path = Path.Relative_path
-module Fpart = Fpart
+module Fsegment = Fsegment
+module Fpart = Fsegment